refactor(host): split BLE/Classic handlers and pairing into separate modules
host.py shrinks from ~1500 to ~650 lines. Per-protocol logic now lives
in dedicated modules:

- ble.py — BLE accept-list / scan handlers, GATT discovery, report
  subscription, HID activation
- classic.py — Classic connection handler, page-scan setup, SDP query,
  active-connect loop
- pairing.py — AutoAcceptPairingDelegate, create_pairing_config,
  create_keystore

HIDHost retains orchestration (run loop, protocol selection,
connection state, UHID device creation, cleanup) and pulls handler
behaviour from the new modules.
